Revert to using doubles for robustness

+++ libraries/Canvas/include/Canvas.hpp
@@ -21,8 +21,8 @@ public:
 private:
   void PPMHeader(std::ostream& outputStream) const noexcept;
   void PPMData(std::ostream& outputStream) const noexcept;
-  inline unsigned int convertColor(const float& colorComponent) const noexcept;
-  size_t ColorComponentToPPM(const float& colorComponent, std::ostream& outputStream, size_t rowLineLen) const noexcept;
+  inline unsigned int convertColor(const double& colorComponent) const noexcept;
+  size_t ColorComponentToPPM(const double& colorComponent, std::ostream& outputStream, size_t rowLineLen) const noexcept;
   size_t pixelToPPM(utility::Color const& pixel, std::ostream& outputStream, size_t rowLineLen) const noexcept;
   std::string rowToPPM(size_t rowIdx) const noexcept;
