+++ libraries/Geometry/include/Cone.hpp
@@ -14,13 +14,13 @@ class Cone final : public ShapeBase
 {
 public:
   Cone() noexcept;
-  Cone(float minimum, float maximum, bool closed = false) noexcept;
+  Cone(double minimum, double maximum, bool closed = false) noexcept;
 
   utility::Tuple localNormalAt(const utility::Tuple &objectPoint) const noexcept;
   std::vector<Intersection> localIntersect(const utility::Ray &transformedRay) const noexcept;
 
-  float minimum_{-std::numeric_limits<float>::infinity()};
-  float maximum_{std::numeric_limits<float>::infinity()};
+  double minimum_{-std::numeric_limits<double>::infinity()};
+  double maximum_{std::numeric_limits<double>::infinity()};
   bool closed_{false};
 };
